FIX AWS

+++ Jenkinsfile
@@ -2,7 +2,7 @@ pipeline {
     agent any
     environment {
         SERVER_REG = "balgittuber"
-        AWS_SERVER = "ubuntu@172.31.20.210"
+        AWS_SERVER = "ubuntu@34.235.175.49"
         AWS_PEM = "/var/jenkins_home/aws/santi.pem"
 
         APP_NAME = "corven-talk-ai"
@@ -22,15 +22,18 @@ pipeline {
                     docker login 
                     docker push ${SERVER_REG}/${APP_NAME}:${BRANCH_NAME}-${BUILD_ID}
 
-                    ssh -i ${AWS_PEM} ${AWS_SERVER} "
-                    docker stop ${APP_NAME}
-                    docker rm ${APP_NAME}
-                    "
+                    ssh -i ${AWS_PEM} ${AWS_SERVER} \"
+                    if docker ps -a --format '{{.Names}}' | grep -Eq '^${APP_NAME}\$'; then
+                        echo 'Stopping and removing existing container: ${APP_NAME}'
+                        docker stop ${APP_NAME}
+                        docker rm ${APP_NAME}
+                    fi
+                    \"
 
-                    ssh -i ${AWS_PEM} ${AWS_SERVER} "
+                    ssh -i ${AWS_PEM} ${AWS_SERVER} \"
                     # Run the new container
                     docker run -d --env-file ${ENV} --name ${APP_NAME} --restart always -p ${PORT}:3000 ${SERVER_REG}/${APP_NAME}:${BRANCH_NAME}-${BUILD_ID}
-                    "
+                    \"
                     """
             }
         }
